// util/types/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Capsule = {
  __typename?: "Capsule";
  /** @deprecated This is not available in the REST API after MongoDB has been deprecated */
  dragon?: Maybe<Dragon>;
  id?: Maybe<Scalars["ID"]["output"]>;
  landings?: Maybe<Scalars["Int"]["output"]>;
  missions?: Maybe<Array<Maybe<CapsuleMission>>>;
  original_launch?: Maybe<Scalars["Date"]["output"]>;
  reuse_count?: Maybe<Scalars["Int"]["output"]>;
  status?: Maybe<Scalars["String"]["output"]>;
  type?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Core = {
  __typename?: "Core";
  asds_attempts?: Maybe<Scalars["Int"]["output"]>;
  asds_landings?: Maybe<Scalars["Int"]["output"]>;
  block?: Maybe<Scalars["Int"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  missions?: Maybe<Array<Maybe<CapsuleMission>>>;
  original_launch?: Maybe<Scalars["Date"]["output"]>;
  reuse_count?: Maybe<Scalars["Int"]["output"]>;
  rtls_attempts?: Maybe<Scalars["Int"]["output"]>;
  rtls_landings?: Maybe<Scalars["Int"]["output"]>;
  status?: Maybe<Scalars["String"]["output"]>;
  water_landing?: Maybe<Scalars["Boolean"]["output"]>;
};
// ...
export type Dragon = {
  __typename?: "Dragon";
  active?: Maybe<Scalars["Boolean"]["output"]>;
  crew_capacity?: Maybe<Scalars["Int"]["output"]>;
  description?: Maybe<Scalars["String"]["output"]>;
  diameter?: Maybe<Distance>;
  dry_mass_kg?: Maybe<Scalars["Int"]["output"]>;
  dry_mass_lb?: Maybe<Scalars["Int"]["output"]>;
  first_flight?: Maybe<Scalars["String"]["output"]>;
  heat_shield?: Maybe<DragonHeatShield>;
  height_w_trunk?: Maybe<Distance>;
  id?: Maybe<Scalars["ID"]["output"]>;
  launch_payload_mass?: Maybe<Mass>;
  launch_payload_vol?: Maybe<Volume>;
  name?: Maybe<Scalars["String"]["output"]>;
  orbit_duration_yr?: Maybe<Scalars["Int"]["output"]>;
  pressurized_capsule?: Maybe<DragonPressurizedCapsule>;
  return_payload_mass?: Maybe<Mass>;
  return_payload_vol?: Maybe<Volume>;
  sidewall_angle_deg?: Maybe<Scalars["Float"]["output"]>;
  thrusters?: Maybe<Array<Maybe<DragonThrust>>>;
  trunk?: Maybe<DragonTrunk>;
  type?: Maybe<Scalars["String"]["output"]>;
  wikipedia?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type HistoriesResult = {
  __typename?: "HistoriesResult";
  data?: Maybe<Array<Maybe<History>>>;
  result?: Maybe<Result>;
};

export type History = {
  __typename?: "History";
  details?: Maybe<Scalars["String"]["output"]>;
  event_date_unix?: Maybe<Scalars["Date"]["output"]>;
  event_date_utc?: Maybe<Scalars["Date"]["output"]>;
  flight?: Maybe<Launch>;
  id?: Maybe<Scalars["ID"]["output"]>;
  links?: Maybe<Link>;
  title?: Maybe<Scalars["String"]["output"]>;
};

export type HistoryFind = {
  end?: InputMaybe<Scalars["Date"]["input"]>;
  flight_number?: InputMaybe<Scalars["Int"]["input"]>;
  id?: InputMaybe<Scalars["ID"]["input"]>;
  start?: InputMaybe<Scalars["Date"]["input"]>;
};

export type Info = {
  __typename?: "Info";
  ceo?: Maybe<Scalars["String"]["output"]>;
  coo?: Maybe<Scalars["String"]["output"]>;
  cto?: Maybe<Scalars["String"]["output"]>;
  cto_propulsion?: Maybe<Scalars["String"]["output"]>;
  employees?: Maybe<Scalars["Int"]["output"]>;
  founded?: Maybe<Scalars["Int"]["output"]>;
  founder?: Maybe<Scalars["String"]["output"]>;
  headquarters?: Maybe<Address>;
  launch_sites?: Maybe<Scalars["Int"]["output"]>;
  links?: Maybe<InfoLinks>;
  name?: Maybe<Scalars["String"]["output"]>;
  summary?: Maybe<Scalars["String"]["output"]>;
  test_sites?: Maybe<Scalars["Int"]["output"]>;
  valuation?: Maybe<Scalars["Float"]["output"]>;
  vehicles?: Maybe<Scalars["Int"]["output"]>;
};

export type InfoLinks = {
  __typename?: "InfoLinks";
  elon_twitter?: Maybe<Scalars["String"]["output"]>;
  flickr?: Maybe<Scalars["String"]["output"]>;
  twitter?: Maybe<Scalars["String"]["output"]>;
  website?: Maybe<Scalars["String"]["output"]>;
};

export type Landpad = {
  __typename?: "Landpad";
  attempted_landings?: Maybe<Scalars["String"]["output"]>;
  details?: Maybe<Scalars["String"]["output"]>;
  full_name?: Maybe<Scalars["String"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  landing_type?: Maybe<Scalars["String"]["output"]>;
  location?: Maybe<Location>;
  status?: Maybe<Scalars["String"]["output"]>;
  successful_landings?: Maybe<Scalars["String"]["output"]>;
  wikipedia?: Maybe<Scalars["String"]["output"]>;
};

export type Launch = {
  __typename?: "Launch";
  details?: Maybe<Scalars["String"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  is_tentative?: Maybe<Scalars["Boolean"]["output"]>;
  launch_date_local?: Maybe<Scalars["Date"]["output"]>;
  launch_date_unix?: Maybe<Scalars["Date"]["output"]>;
  launch_date_utc?: Maybe<Scalars["Date"]["output"]>;
  launch_site?: Maybe<LaunchSite>;
  launch_success?: Maybe<Scalars["Boolean"]["output"]>;
  launch_year?: Maybe<Scalars["String"]["output"]>;
  links?: Maybe<LaunchLinks>;
  mission_id?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  mission_name?: Maybe<Scalars["String"]["output"]>;
  rocket?: Maybe<LaunchRocket>;
  ships?: Maybe<Array<Maybe<Ship>>>;
  static_fire_date_unix?: Maybe<Scalars["Date"]["output"]>;
  static_fire_date_utc?: Maybe<Scalars["Date"]["output"]>;
  telemetry?: Maybe<LaunchTelemetry>;
  tentative_max_precision?: Maybe<Scalars["String"]["output"]>;
  upcoming?: Maybe<Scalars["Boolean"]["output"]>;
};
// ...
export type LaunchesPastResult = {
  __typename?: "LaunchesPastResult";
  data?: Maybe<Array<Maybe<Launch>>>;
  result?: Maybe<Result>;
};

export type Launchpad = {
  __typename?: "Launchpad";
  attempted_launches?: Maybe<Scalars["Int"]["output"]>;
  details?: Maybe<Scalars["String"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  location?: Maybe<Location>;
  name?: Maybe<Scalars["String"]["output"]>;
  status?: Maybe<Scalars["String"]["output"]>;
  successful_launches?: Maybe<Scalars["Int"]["output"]>;
  vehicles_launched?: Maybe<Array<Maybe<Rocket>>>;
  wikipedia?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Mission = {
  __typename?: "Mission";
  description?: Maybe<Scalars["String"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  manufacturers?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  name?: Maybe<Scalars["String"]["output"]>;
  payloads?: Maybe<Array<Maybe<Payload>>>;
  twitter?: Maybe<Scalars["String"]["output"]>;
  website?: Maybe<Scalars["String"]["output"]>;
  wikipedia?: Maybe<Scalars["String"]["output"]>;
};

export type MissionResult = {
  __typename?: "MissionResult";
  data?: Maybe<Array<Maybe<Mission>>>;
  result?: Maybe<Result>;
};
// ...
export type Payload = {
  __typename?: "Payload";
  customers?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  id?: Maybe<Scalars["ID"]["output"]>;
  manufacturer?: Maybe<Scalars["String"]["output"]>;
  nationality?: Maybe<Scalars["String"]["output"]>;
  norad_id?: Maybe<Array<Maybe<Scalars["Int"]["output"]>>>;
  orbit?: Maybe<Scalars["String"]["output"]>;
  orbit_params?: Maybe<PayloadOrbitParams>;
  payload_mass_kg?: Maybe<Scalars["Float"]["output"]>;
  payload_mass_lbs?: Maybe<Scalars["Float"]["output"]>;
  payload_type?: Maybe<Scalars["String"]["output"]>;
  reused?: Maybe<Scalars["Boolean"]["output"]>;
};
// ...
export type Query = {
  __typename?: "Query";
  _service: _Service;
  capsule?: Maybe<Capsule>;
  capsules?: Maybe<Array<Maybe<Capsule>>>;
  capsulesPast?: Maybe<Array<Maybe<Capsule>>>;
  capsulesUpcoming?: Maybe<Array<Maybe<Capsule>>>;
  company?: Maybe<Info>;
  core?: Maybe<Core>;
  cores?: Maybe<Array<Maybe<Core>>>;
  coresPast?: Maybe<Array<Maybe<Core>>>;
  coresUpcoming?: Maybe<Array<Maybe<Core>>>;
  dragon?: Maybe<Dragon>;
  dragons?: Maybe<Array<Maybe<Dragon>>>;
  histories?: Maybe<Array<Maybe<History>>>;
  historiesResult?: Maybe<HistoriesResult>;
  history?: Maybe<History>;
  landpad?: Maybe<Landpad>;
  landpads?: Maybe<Array<Maybe<Landpad>>>;
  launch?: Maybe<Launch>;
  launchLatest?: Maybe<Launch>;
  launchNext?: Maybe<Launch>;
  launches?: Maybe<Array<Maybe<Launch>>>;
  launchesPast?: Maybe<Array<Maybe<Launch>>>;
  launchesPastResult?: Maybe<LaunchesPastResult>;
  launchesUpcoming?: Maybe<Array<Maybe<Launch>>>;
  launchpad?: Maybe<Launchpad>;
  launchpads?: Maybe<Array<Maybe<Launchpad>>>;
  /** @deprecated Mission is not available on REST API after MongoDB deprecation */
  mission?: Maybe<Mission>;
  /** @deprecated Mission is not available on REST API after MongoDB deprecation */
  missions?: Maybe<Array<Maybe<Mission>>>;
  /** @deprecated Mission is not available on REST API after MongoDB deprecation */
  missionsResult?: Maybe<MissionResult>;
  payload?: Maybe<Payload>;
  payloads?: Maybe<Array<Maybe<Payload>>>;
  roadster?: Maybe<Roadster>;
  rocket?: Maybe<Rocket>;
  rockets?: Maybe<Array<Maybe<Rocket>>>;
  rocketsResult?: Maybe<RocketsResult>;
  ship?: Maybe<Ship>;
  ships?: Maybe<Array<Maybe<Ship>>>;
  shipsResult?: Maybe<ShipsResult>;
  /** fetch data from the table: "users" */
  users: Array<Users>;
  /** fetch aggregated fields from the table: "users" */
  users_aggregate: Users_Aggregate;
  /** fetch data from the table: "users" using primary key columns */
  users_by_pk?: Maybe<Users>;
};
// ...
export type Roadster = {
  __typename?: "Roadster";
  apoapsis_au?: Maybe<Scalars["Float"]["output"]>;
  details?: Maybe<Scalars["String"]["output"]>;
  earth_distance_km?: Maybe<Scalars["Float"]["output"]>;
  earth_distance_mi?: Maybe<Scalars["Float"]["output"]>;
  eccentricity?: Maybe<Scalars["Float"]["output"]>;
  epoch_jd?: Maybe<Scalars["Float"]["output"]>;
  inclination?: Maybe<Scalars["Float"]["output"]>;
  launch_date_unix?: Maybe<Scalars["Date"]["output"]>;
  launch_date_utc?: Maybe<Scalars["Date"]["output"]>;
  launch_mass_kg?: Maybe<Scalars["Int"]["output"]>;
  launch_mass_lbs?: Maybe<Scalars["Int"]["output"]>;
  longitude?: Maybe<Scalars["Float"]["output"]>;
  mars_distance_km?: Maybe<Scalars["Float"]["output"]>;
  mars_distance_mi?: Maybe<Scalars["Float"]["output"]>;
  name?: Maybe<Scalars["String"]["output"]>;
  norad_id?: Maybe<Scalars["Int"]["output"]>;
  orbit_type?: Maybe<Scalars["Float"]["output"]>;
  periapsis_arg?: Maybe<Scalars["Float"]["output"]>;
  periapsis_au?: Maybe<Scalars["Float"]["output"]>;
  period_days?: Maybe<Scalars["Float"]["output"]>;
  semi_major_axis_au?: Maybe<Scalars["Float"]["output"]>;
  speed_kph?: Maybe<Scalars["Float"]["output"]>;
  speed_mph?: Maybe<Scalars["Float"]["output"]>;
  wikipedia?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type RocketsResult = {
  __typename?: "RocketsResult";
  data?: Maybe<Array<Maybe<Rocket>>>;
  result?: Maybe<Result>;
};

export type Ship = {
  __typename?: "Ship";
  abs?: Maybe<Scalars["Int"]["output"]>;
  active?: Maybe<Scalars["Boolean"]["output"]>;
  attempted_landings?: Maybe<Scalars["Int"]["output"]>;
  class?: Maybe<Scalars["Int"]["output"]>;
  course_deg?: Maybe<Scalars["Int"]["output"]>;
  home_port?: Maybe<Scalars["String"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  image?: Maybe<Scalars["String"]["output"]>;
  imo?: Maybe<Scalars["Int"]["output"]>;
  missions?: Maybe<Array<Maybe<ShipMission>>>;
  mmsi?: Maybe<Scalars["Int"]["output"]>;
  model?: Maybe<Scalars["String"]["output"]>;
  name?: Maybe<Scalars["String"]["output"]>;
  position?: Maybe<ShipLocation>;
  roles?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>;
  speed_kn?: Maybe<Scalars["Float"]["output"]>;
  status?: Maybe<Scalars["String"]["output"]>;
  successful_landings?: Maybe<Scalars["Int"]["output"]>;
  type?: Maybe<Scalars["String"]["output"]>;
  url?: Maybe<Scalars["String"]["output"]>;
  weight_kg?: Maybe<Scalars["Int"]["output"]>;
  weight_lbs?: Maybe<Scalars["Int"]["output"]>;
  year_built?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type _Service = {
  __typename?: "_Service";
  sdl?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type ShipsQuery = {
  __typename?: "Query";
  ships?: Array<{
    __typename?: "Ship";
    id?: string | null;
    image?: string | null;
    class?: number | null;
    active?: boolean | null;
    name?: string | null;
  } | null> | null;
};
// ...
export type MissionsQueryVariables = Exact<{
  shipId: string;
}>;

export type MissionsQuery = {
  __typename?: "Query";
  missions?: Array<{
    __typename?: "Mission";
    id?: string | null;
    active?: boolean | null;
    cargo?: string | null;
    destination?: string | null;
    name?: string | null;
    shipId?: string | null;
  } | null> | null;
};